Add conv2d with shared memory
This is a new version of conv2d with shared memory.
In this version, we support any filter size, padding,
stride and dilation, and it works well on mobilenet.

But there are some limitations to output texture shape:
1. Output texture shape must be [NHW, C]
2. OutWidth should be divisible by localGroupSize[1]

TODO:
1. Use Conv2DProgram if tensor size is not large enough
2. Output texture shape can be [N, HWC]

Besides, we define a function getFirstThreadOutputCoords()
which returns the 4d coordinate of first thread in local
group.